+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 1. nginx 删除操作新增还原配置
 2. 【server】新增用户登录日志记录
 3. 【server】log view 新增正则参考
+4. 【server】docker 控制台新增 docker 信息查看
+5. 【server】新增 docker swarm 集群管理
 
 ### 🐞 解决BUG、优化功能